remove the float to half later

+++ CAESAR/models/caesar_compress.cpp
@@ -321,7 +321,7 @@ CompressionResult Compressor::compress(const DatasetConfig& config , int batch_s
 
             torch::Tensor batched_indexes = torch::cat(batch_indexes , 0).to(device_);
 
-            std::vector<torch::Tensor> inputs = { batched_input.to(torch::kFloat16) };
+            std::vector<torch::Tensor> inputs = { batched_input.to(torch::kFloat32) };
             std::vector<torch::Tensor> outputs = compressor_model_->run(inputs);
 
             torch::Tensor latent = outputs[0];
@@ -331,10 +331,10 @@ CompressionResult Compressor::compress(const DatasetConfig& config , int batch_s
             outputs.shrink_to_fit();
 
             std::vector<torch::Tensor> hyper_outputs = hyper_decompressor_model_->run({ q_hyper_latent.to(torch::kDouble) });
-torch::Tensor mean = hyper_outputs[0].to(torch::kFloat32);
-torch::Tensor latent_indexes_recon = hyper_outputs[1].to(torch::kFloat32);
-hyper_outputs.clear();
-hyper_outputs.shrink_to_fit();
+            torch::Tensor mean = hyper_outputs[0].to(torch::kFloat32);
+            torch::Tensor latent_indexes_recon = hyper_outputs[1].to(torch::kFloat32);
+            hyper_outputs.clear();
+            hyper_outputs.shrink_to_fit();
 
 
             hyper_outputs.clear();
